pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcpbr"
version = "0.3.1"
description = "Model Context Protocol Benchmark Runner - evaluate MCP servers against software engineering benchmarks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "mcpbr Contributors" }
]
keywords = ["mcp", "model-context-protocol", "swe-bench", "cybergym", "benchmark", "llm", "agents", "evaluation", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "anthropic>=0.40.0",
    "mcp>=1.0.0",
    "docker>=7.0.0",
    "datasets>=2.14.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-minify-plugin>=0.7.0",
]
# Note: terminal-bench package (pip install terminal-bench) can be installed
# separately if HuggingFace dataset loading fails. It requires Python 3.12+.

[project.scripts]
mcpbr = "mcpbr.cli:main"

[project.urls]
Homepage = "https://github.com/greynewell/mcpbr"
Repository = "https://github.com/greynewell/mcpbr"
Documentation = "https://greynewell.github.io/mcpbr/"
Changelog = "https://github.com/greynewell/mcpbr/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/greynewell/mcpbr/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/mcpbr"]

[tool.hatch.build.targets.wheel.shared-data]
"src/mcpbr/data" = "mcpbr/data"

[tool.hatch.build.targets.sdist]
exclude = [
    "docs/",
    "mkdocs.yml",
    "site/",
    ".github/workflows/docs.yml",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m not integration')",
]
